+++ src/domain/application/services/user/delete-user.ts
@@ -1,5 +1,7 @@
 import { Either, left, right } from '@/@shared/either'
+import { Injectable } from '@nestjs/common'
 import { CacheRepository } from '../../cache/cache-repository'
+import { InactiveUserError } from '../../errors/InactiveUserError'
 import { UserNonExistsError } from '../../errors/UserNonExists'
 import { UserRepository } from '../../repositories/user-repository'
 
@@ -7,8 +9,12 @@ type DeleteUserServiceRequest = {
   id: string
 }
 
-type DeleteUserServiceResponse = Either<UserNonExistsError, boolean>
+type DeleteUserServiceResponse = Either<
+  UserNonExistsError | InactiveUserError,
+  boolean
+>
 
+@Injectable()
 export class DeleteUserService {
   constructor(
     private userRepository: UserRepository,
@@ -22,6 +28,10 @@ export class DeleteUserService {
 
     if (!user) {
       return left(new UserNonExistsError())
+    }
+
+    if (!user.active) {
+      return left(new InactiveUserError())
     }
 
     const result = await this.userRepository.deleteUser(id)
